chore(backend): moved /:userId/drafts endpoint to the UsersController
