+++ apps/zero-api/src/users/users.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { UsersDraftsController } from './users-drafts.controller';
 import { UsersFilesController } from './users-files.controller';
 import { DraftsModule } from '../drafts/drafts.module';
 import { FilesModule } from '../files/files.module';
 
 @Module({
-  controllers: [UsersController, UsersDraftsController, UsersFilesController],
+  controllers: [UsersController, UsersFilesController],
   imports: [DraftsModule, FilesModule],
   providers: [UsersService],
   exports: [UsersService]
